CR-71764,CR-71066: Fix kernel panic resulting from linux upgrade
Commit c0d59da79534 ("ip_gre: Make none-tun-dst gre tunnel store tunnel
info as metadat_dst in recv") changed it so that packets with no
destination also store the metadata_dst meaning that the destination
pointer on the socket buffer will always be allocated for packets over
GRE tunnels. By default the metadata destination has a device pointer
set to NULL.

Code in ipt_netmap.c assumed that if the destination is set then we will
also have set the device pointer which results in a crash when we try to
access the MTU field of the destination device.

Updated this condition statement to also make sure the destination
device is not NULL before trying to use the ip_skb_dst_mtu function.

Change-Id: I2376d3f3a5aaac80faea075e28562664856a10a5
Reviewed-by: Carl Smith <carl.smith@alliedtelesis.co.nz>
Reviewed-by: Blair Steven <blair.steven@alliedtelesis.co.nz>