fix copyin/out of nmreq option list
The previous code tried unsuccesfully to report a precise error for each
option in the user list. Moreover, commit 253b2ec199b broke some
ctrl-api-test (see
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=260547).

The new code bails out as soon as an unrecoverable error is detected and
properly checks for copy boundaries. EOPNOTSUPP no longer immediately
returns an error, so that any other option in the list may be examined
by the caller code and a precise report of the (un)supported options can
be returned to the user.

With this patch, all ctrl-api-test unit tests can be passed again.