Process pkts on netmap user thread/stack rather than ksoftirqd's
The advantage of ksoftirqd is that it's threads can be
prioitised against all other threads in the system and that
it has a known stack size/state. We get the same benefits from
processing packets on the netmap user thread/stack without
queuing to the cpu backlog and context switching to the
ksoftirqd process.