mvpp2: set rx offset using queue id
If multiple ports are used on the same CP unit, the queue number for
these are separated by 32 i.e. queue 0 on port 0 is q0, but queue 0 on
port 1 is q32. Therefore the offset set code does not work for these
ports, as it tries to set q0-3 on each port instead of the correct queue
ids.